# specforge-openapi.yaml
openapi: 3.1.0
info:
  title: SpecForge API
  version: 1.0.0
  description: |
    SpecForge is a spec-first roadmap and contract governance platform
    designed for AI-assisted software development.
servers:
  - url: https://api.specforge.io/api/v1
security:
  - bearerAuth: []
...
paths:
...
  /import/init:
    post:
      tags: [Import]
      summary: Initialize existing project import session
      operationId: initProjectImport
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImportInitRequest'
      responses:
        "200":
          description: Import session initialized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportInitResponse'
  /import/snapshot:
    post:
      tags: [Import]
      summary: Submit project reality snapshot
      operationId: submitProjectSnapshot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectSnapshotSubmission'
      responses:
        "200":
          description: Snapshot analyzed and scored
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SnapshotAnalysisResponse'
  /import/alignment-rules:
    get:
      tags: [Import]
      summary: Retrieve strict alignment rules after import
      operationId: getImportAlignmentRules
      parameters:
        - in: query
          name: project_id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Strict IDE enforcement rules
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlignmentRulesResponse'
  /import/post-snapshot:
    post:
      tags: [Import]
      summary: Submit post-implementation snapshot for drift verification
      operationId: submitPostImportSnapshot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostImportSnapshotSubmission'
      responses:
        "200":
          description: Drift analysis completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostSnapshotAnalysisResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    ImportInitRequest:
      type: object
      required:
        - project_id
        - language_stack
        - repository_type
        - estimated_size
        - import_mode
      properties:
        project_id:
          type: string
        language_stack:
          type: array
          items:
            type: string
        repository_type:
          type: string
          enum: [monorepo, polyrepo, single]
        estimated_size:
          type: string
          enum: [small, medium, large]
        import_mode:
          type: string
          enum: [light, full]
    ImportInitResponse:
      type: object
      properties:
        session_id:
          type: string
        extraction_contract:
          type: object
          description: Required structured snapshot schema reference
        validation_rules:
          type: array
          items:
            type: string
        max_payload_size_mb:
          type: integer

    ########################
    # SNAPSHOT SUBMISSION
    ########################
    ProjectSnapshotSubmission:
      type: object
      required:
        - project_id
        - snapshot_version
        - snapshot_payload
      properties:
        project_id:
          type: string
        snapshot_version:
          type: string
        snapshot_payload:
          $ref: '#/components/schemas/ProjectSnapshot'
    ProjectSnapshot:
      type: object
      required:
        - metadata
        - architecture
        - contracts
        - variables
        - security
        - tests
      properties:
        metadata:
          $ref: '#/components/schemas/ProjectMetadata'
        architecture:
          $ref: '#/components/schemas/ArchitectureModel'
        contracts:
          $ref: '#/components/schemas/ContractsModel'
        variables:
          $ref: '#/components/schemas/VariablesModel'
        security:
          $ref: '#/components/schemas/SecurityModel'
        tests:
          $ref: '#/components/schemas/TestModel'
    ProjectMetadata:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        languages:
          type: array
          items:
            type: string
        frameworks:
          type: array
          items:
            type: string
        entrypoints:
          type: array
          items:
            type: string
        runtime_type:
          type: string
        deployment_type:
          type: string
    ArchitectureModel:
      type: object
      properties:
        layers:
          type: array
          items:
            type: string
        service_boundaries:
          type: array
          items:
            type: string
        dependencies_graph:
          type: array
          items:
            $ref: '#/components/schemas/DependencyEdge'
    DependencyEdge:
      type: object
      required: [from, to]
      properties:
        from:
          type: string
        to:
          type: string
        type:
          type: string
    ContractsModel:
      type: object
      properties:
        api_contracts:
          type: array
          items:
            $ref: '#/components/schemas/APIContract'
        data_models:
          type: array
          items:
            $ref: '#/components/schemas/DataModel'
        validation_rules:
          type: array
          items:
            type: string
    APIContract:
      type: object
      required: [name, method, path]
      properties:
        name:
          type: string
        method:
          type: string
        path:
          type: string
        request_schema:
          type: object
        response_schema:
          type: object
    DataModel:
      type: object
      required: [name]
      properties:
        name:
          type: string
        fields:
          type: array
          items:
            $ref: '#/components/schemas/DataField'
    DataField:
      type: object
      required: [name, type]
      properties:
        name:
          type: string
        type:
          type: string
        required:
          type: boolean
    VariablesModel:
      type: object
      properties:
        global_config:
          type: array
          items:
            type: string
        environment_variables:
          type: array
          items:
            type: string
        feature_flags:
          type: array
          items:
            type: string
    SecurityModel:
      type: object
      properties:
        auth_mechanisms:
          type: array
          items:
            type: string
        roles:
          type: array
          items:
            type: string
        external_integrations:
          type: array
          items:
            type: string
    TestModel:
      type: object
      properties:
        coverage_estimate:
          type: number
          minimum: 0
          maximum: 100
        test_types:
          type: array
          items:
            type: string

    ########################
    # SNAPSHOT ANALYSIS RESPONSE
    ########################
    SnapshotAnalysisResponse:
      type: object
      properties:
        alignment_score:
          type: integer
          minimum: 0
          maximum: 100
        architectural_risk_score:
          type: integer
        coupling_score:
          type: integer
        contract_integrity_score:
          type: integer
        missing_intelligence_flags:
          type: array
          items:
            type: string
        required_next_action:
          type: string

    ########################
    # ALIGNMENT RULES
    ########################
    AlignmentRulesResponse:
      type: object
      properties:
        strict_rules:
          type: array
          items:
            type: string
        forbidden_actions:
          type: array
          items:
            type: string
        required_snapshot_policy:
          type: object
          properties:
            must_create_snapshot_before_changes:
              type: boolean
            must_post_snapshot_after_changes:
              type: boolean
        alignment_constraints:
          type: object
          additionalProperties:
            type: boolean

    ########################
    # POST SNAPSHOT
    ########################
    PostImportSnapshotSubmission:
      type: object
      required:
        - project_id
        - snapshot_version
        - snapshot_payload
      properties:
        project_id:
          type: string
        snapshot_version:
          type: string
        snapshot_payload:
          $ref: '#/components/schemas/ProjectSnapshot'
    PostSnapshotAnalysisResponse:
      type: object
      properties:
        drift_score:
          type: integer
          minimum: 0
          maximum: 100
        contract_breakage_detected:
          type: boolean
        risk_delta:
          type: number
        alignment_delta:
          type: number
        recommendation:
          type: string
          enum: [approve, revise, reject]
...